Ready for todos/:id

+++ backend/internal/service/delete_task.go
@@ -1,9 +1,19 @@
 package service
 
-import "context"
+import (
+	"context"
+
+	errservice "github.com/federalbyro/to-do-app/backend/internal/service/errors"
+)
 
 func (s *TaskService) DeleteTask(c context.Context, user_id, id string) error {
 	key := user_id + ":" + id
+	_, exist := s.RedisCache.GetTaskRedis(c, key)
+
+	if exist != nil {
+		return errservice.ErrExistBase
+	}
+
 	err := s.RedisCache.DeleteTaskRedis(c, key)
 	return err
 }
